requirements.txt
-e git://github.com/defrex/django-security-middleware.git#egg=django-security-middleware
-e git://github.com/defrex/django-bootstrap-paginator.git#egg=django-bootstrap-paginator
-e git://github.com/defrex/django-resize#egg=django-resize

Django==1.9.1
psycopg2==2.6.1
django-filter==0.12.0
django-bootstrap-form==3.2
python-dateutil==2.4.2
factory-boy==2.8.1
Delorean==0.6.0
tqdm==3.8.0
django-watchman==0.9.0
django-extensions==1.6.1
requests==2.7.0
xmltodict==0.10.1
Pillow==3.2.0
raven==5.12.0
markdown==2.6.6
tweepy==3.5.0
django-csp==2.0.3
graphql-core==0.5.2
graphql-django-view==1.3
graphql-relay==0.4.2
graphene==0.10.2
django-graphiql==0.4.4
html2text==2016.5.29
geoip2==2.3.0
django-ipware==1.1.5
mixpanel==4.3.0
celery==3.1.23
redis==2.10.5
django-redis==4.4.3
django-celery==3.1.17
django-markdown==0.8.4
django-taggit==0.18.1
premailer==3.0.1
django-mathfilters==0.4.0
